+++ management/model_image.go
@@ -19,6 +19,8 @@ type Image struct {
 	// A string that specifies the resource’s unique identifier.
 	Id *string `json:"id,omitempty"`
 	Environment *ObjectEnvironment `json:"environment,omitempty"`
+	// The time the resource was created.
+	CreatedAt *string `json:"createdAt,omitempty"`
 	Targets *ImageTargets `json:"targets,omitempty"`
 }
 
@@ -103,6 +105,38 @@ func (o *Image) SetEnvironment(v ObjectEnvironment) {
 	o.Environment = &v
 }
 
+// GetCreatedAt returns the CreatedAt field value if set, zero value otherwise.
+func (o *Image) GetCreatedAt() string {
+	if o == nil || o.CreatedAt == nil {
+		var ret string
+		return ret
+	}
+	return *o.CreatedAt
+}
+
+// GetCreatedAtOk returns a tuple with the CreatedAt field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *Image) GetCreatedAtOk() (*string, bool) {
+	if o == nil || o.CreatedAt == nil {
+		return nil, false
+	}
+	return o.CreatedAt, true
+}
+
+// HasCreatedAt returns a boolean if a field has been set.
+func (o *Image) HasCreatedAt() bool {
+	if o != nil && o.CreatedAt != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetCreatedAt gets a reference to the given string and assigns it to the CreatedAt field.
+func (o *Image) SetCreatedAt(v string) {
+	o.CreatedAt = &v
+}
+
 // GetTargets returns the Targets field value if set, zero value otherwise.
 func (o *Image) GetTargets() ImageTargets {
 	if o == nil || o.Targets == nil {
@@ -142,6 +176,9 @@ func (o Image) MarshalJSON() ([]byte, error) {
 	}
 	if o.Environment != nil {
 		toSerialize["environment"] = o.Environment
+	}
+	if o.CreatedAt != nil {
+		toSerialize["createdAt"] = o.CreatedAt
 	}
 	if o.Targets != nil {
 		toSerialize["targets"] = o.Targets